Apply input resampling for motion events of ToolType::MOUSE and STYLUS
As the resampling is performed for better scrolling, even when a user
is dragging with a mouse or a stylus, we'd like to do the same as
touch inputs.

Also, when an event being resampled has any pointer with a tool type
that we don’t resample, the resampled event has an old coordinate but
a new timestamp. This event is confusing. With this change, when the
event has such a pointer, events won’t be resampled.

Bug: 337849873
Test: Enable debug and resampling happens for mouse events.
Test: TouchResamplingTest
Change-Id: I9344759e28685a08e233939761f93f0b3911413c