Pass state instead of unsafe display on binder
Accessing display from pickBestDataspace may be unsafe
without mStateLock. The display was originally used to
get the OutputCompositionState, which can instead be
passed in directly. There is no need to check for a null
display in this method anymore because the caller of this
function already has the correct default value.

Bug: b/294936197
Test: presubmit
Change-Id: I10ce68a75f91d14fb18f457ac4eebb2ed54fe09a