Verify all injected events
Prior to this CL, we only verified the streams from events injected from
accessibility.

With this CL, we will now verify all injected events, even if they did
not originate from a11y. That means, these events were injected from
tests. For now, we will only log the fact that the stream is
inconsistent, but still let it through. Eventually, we can fail the
injection instead, which will prevent crashes in the dispatcher.

Bug: 329439551
Test: atest inputflinger_tests
Change-Id: Ie7313fd79f614d3347ef7198812be3c810eb86ea