InputDispatcher: Fix multi-display Pointer Capture
There is an existing requirement that a window must both have focus and
be on the focused display to be able to gain Pointer Capture.

This means that focus changes on non-focused displays should not affect
Pointer Capture, and that a window must lose capture if its display
loses focus.

Verify these requirements with a test.

Bug: 342229227
Test: atest inputflinger_tests
Change-Id: I7b1c73b7759d8f20436ee401ba657a5dc8ead7a5